make tests client functional
