+++ sunpy/net/dataretriever/attrs/__init__.py
@@ -1,3 +1,8 @@
 from . import goes
+from . import adapt
 
-__all__ = ['goes']
+__all__ = ['goes', 'adapt']
+
+# Define a custom __dir__ to restrict tab-completion to __all__
+def __dir__():
+    return __all__
