+++ sunpy/net/dataretriever/sources/__init__.py
@@ -1,0 +1,2 @@
+# from . import adapt
+__all__ = ['adapt']
