+++ sunpy/net/dataretriever/sources/tests/test_adapt.py
@@ -1,7 +1,7 @@
 import pytest
 from sunpy.net import attrs
-from sunpy.net.dataretriever.sources.adapt import carrington_time
-from sunpy.net.dataretriever.attrs.adapt import *
+from sunpy.net.dataretriever.sources import adapt as adapt_source
+from sunpy.net.dataretriever.attrs.adapt import ADAPTLngType
 import os
 from sunpy.net.fido_factory import Fido
 
@@ -9,24 +9,25 @@ global result
 result = None
 
 @pytest.mark.remote_data
-def test_search():
-    CR=2193 
-    frames=1
-    test=True
+def test_search(CR=2193, frames=1, test=True):
+
     # Get the date range
-    get_date, get_date_end = carrington_time(CR, frames)
+    get_date, get_date_end = adapt_source.carrington_time(CR, frames)
     LngType = '0' # 0 is carrington, 1 is central meridian
+    print (f"\n\n\n\n\n\n\nSearching for images from {get_date} to {get_date_end}\n")
     res = Fido.search(attrs.Instrument('adapt'), attrs.Time(get_date, get_date_end), ADAPTLngType(LngType))
+    print(res)
     global result
     result = res
     length_test(result, test=test)
-    # print(f"fetch_in = {out}", flush=True)
-    # return out
+    return result
 
 @pytest.mark.remote_data
 def test_fetch(search_result):
+    print(f"Fetching search_result:")
     ret =  Fido.fetch(search_result, path=make_dir())
     length_test(ret, test=True)
+    print("Success!\n\n")
 
 
 @pytest.fixture
@@ -61,36 +62,32 @@ def make_dir(folder="data"):
         os.makedirs(directory)
     return directory
 
-# print(__name__)
+
+def test_client(CR=2193, frames=1, ask=False):
+    res = test_search(CR, frames)
+    # print("Found {} files".format(len(res)))
+    # print(f"Client res is {res}")
+    # ask the user to continue or not
+    assert res
+    if len(res) > 0:
+        if ask:
+            print("Do you want to continue?")
+            inp = input("y/n: ")
+            if inp == 'y':
+                out = test_fetch(res)
+            else:
+                print("Aborting")
+                return None
+        else:
+            out = test_fetch(res)
+    return out
+
+
 if __name__ == "__main__":
-    test_search()
-    test_fetch(search_result=search_result)
-#     # import os
-#     # os.execv('/usr/bin/python3', ['python3'] + sys.argv)
-#     # /opt/homebrew/anaconda3/envs/sunpy-dev/bin/python
-#     pass
-#     # test_client()
+    test_client()
+
 
     # from sunpy.tests.helpers import remote_data
 # from pytest.mark import remote_data
 
 
-# def test_client(CR=2193, frames=1, ask=False):
-#     res = test_search(CR, frames)
-#     # print("Found {} files".format(len(res)))
-#     print(f"Client res is {res}")
-#     # ask the user to continue or not
-#     assert res
-#     if len(res) > 0:
-#         if ask:
-#             print("Do you want to continue?")
-#             inp = input("y/n: ")
-#             if inp == 'y':
-#                 out = test_fetch(res)
-#             else:
-#                 print("Aborting")
-#                 return None
-#         else:
-#             out = test_fetch(res)
-#     print(out)
-#     return out
